+++ apps/ledger-live-desktop/src/renderer/families/tezos/AccountHeaderManageActions.ts
@@ -16,6 +16,7 @@ const AccountHeaderManageActions: TezosFamily["accountHeaderManageActions"] = ({
   const { t } = useTranslation();
   const dispatch = useDispatch();
   const delegation = useDelegation(account);
+
   const onClick = useCallback(() => {
     if (isAccountEmpty(account)) {
       dispatch(
@@ -41,7 +42,9 @@ const AccountHeaderManageActions: TezosFamily["accountHeaderManageActions"] = ({
         };
     dispatch(openModal("MODAL_DELEGATE", options));
   }, [account, delegation, parentAccount, dispatch, source]);
+
   if (parentAccount) return null;
+
   return [
     {
       key: "Stake",