+++ apps/ledger-live-mobile/src/families/evm/EvmFeesStrategy.tsx
@@ -1,18 +1,19 @@
+import { getEstimatedFees } from "@ledgerhq/coin-evm/lib/logic";
+import type { Transaction } from "@ledgerhq/coin-evm/types/index";
 import { getMainAccount } from "@ledgerhq/live-common/account/helpers";
 import { getAccountBridge } from "@ledgerhq/live-common/bridge/index";
-import type { Transaction } from "@ledgerhq/coin-evm/types/index";
-import React, { useCallback, useEffect, useState } from "react";
-import SelectFeesStrategy from "./SelectFeesStrategy";
-import { ScreenName } from "../../const";
-import { EvmNetworkFeeInfo } from "./EvmNetworkFeesInfo";
-import { SendRowsFeeProps as Props } from "./types";
 import { useGasOptions } from "@ledgerhq/live-common/families/evm/react";
 import { log } from "@ledgerhq/logs";
+import { InfiniteLoader } from "@ledgerhq/native-ui";
 import { AccountBridge } from "@ledgerhq/types-live";
 import BigNumber from "bignumber.js";
-import { getEstimatedFees } from "@ledgerhq/coin-evm/lib/logic";
+import React, { useCallback, useEffect, useState } from "react";
+import { ScreenName } from "../../const";
+import { EvmNetworkFeeInfo } from "./EvmNetworkFeesInfo";
+import SelectFeesStrategy from "./SelectFeesStrategy";
+import { SendRowsFeeProps as Props, StrategyWithCustom } from "./types";
 
-const getCustomStrategy = (transaction: Transaction): BigNumber | null => {
+const getCustomStrategyFees = (transaction: Transaction): BigNumber | null => {
   if (transaction.feesStrategy === "custom") {
     return getEstimatedFees(transaction);
   }
@@ -27,12 +28,13 @@ export default function EvmFeesStrategy({
   setTransaction,
   navigation,
   route,
+  shouldPrefillEvmGasOptions = true,
   ...props
 }: Props<Transaction>) {
   const account = getMainAccount(accountProp, parentAccount);
   const bridge: AccountBridge<Transaction> = getAccountBridge(account);
 
-  const [gasOptions, error] = useGasOptions({
+  const [gasOptions, error, loading] = useGasOptions({
     currency: account.currency,
     transaction,
     interval: account.currency.blockAvgTime ? account.currency.blockAvgTime * 1000 : undefined,
@@ -43,33 +45,54 @@ export default function EvmFeesStrategy({
   }
 
   useEffect(() => {
-    const updatedTransaction = bridge.updateTransaction(transaction, {
-      ...transaction,
-      gasOptions,
-    });
-    setTransaction(updatedTransaction);
-  }, [bridge, setTransaction, gasOptions, transaction]);
-
-  const [customStrategy, setCustomStrategy] = useState(getCustomStrategy(transaction));
-
-  useEffect(() => {
-    const newCustomStrategy = getCustomStrategy(transaction);
-
-    if (newCustomStrategy) {
-      setCustomStrategy(newCustomStrategy);
+    if (shouldPrefillEvmGasOptions) {
+      const updatedTransaction = bridge.updateTransaction(transaction, {
+        ...transaction,
+        gasOptions,
+      });
+      setTransaction(updatedTransaction);
     }
-  }, [transaction, setCustomStrategy]);
+  }, [bridge, setTransaction, gasOptions, transaction, shouldPrefillEvmGasOptions]);
+
+  /**
+   * When the user edits the custom fees, we save the related transaction patch
+   * in order to reapply it the next time the user selects the saved custom
+   * strategy entry in the strategy list.
+   * This is done in order to display the correct fees value for the saved custom
+   * entry in the fees strategy list.
+   * We don't need to save the transaction patch for other fees strategies since
+   * the transaction is already updated with the correct fee data retrieved from
+   * the gasTracker in `prepareTransaction`.
+   * cf. (as of 21-09-2023) https://github.com/LedgerHQ/ledger-live/blob/e87c4c2d879fb9322cd2849f2d34c3be9e500a1e/libs/coin-evm/src/prepareTransaction.ts#L190-L202
+   * But since the custom strategy is handled locally by the client, if we don't
+   * save the patch (including updated fee data) and only update the `feesStrategy`
+   * field, the fees value displayed in the fees strategy for the custom entry
+   * will be the one from the last selected strategy (e.g. slow, medium, fast).
+   */
+  const [customStrategyTransactionPatch, setCustomStrategyTransactionPatch] =
+    useState<Partial<Transaction>>();
 
   const onFeesSelected = useCallback(
-    ({ feesStrategy }) => {
-      const bridge = getAccountBridge(account, parentAccount);
-      setTransaction(
-        bridge.updateTransaction(transaction, {
-          feesStrategy,
-        }),
-      );
+    ({ feesStrategy }: { feesStrategy: StrategyWithCustom }) => {
+      const bridge = getAccountBridge<Transaction>(account, parentAccount);
+
+      const patch: Partial<Transaction> =
+        feesStrategy === "custom" && customStrategyTransactionPatch
+          ? customStrategyTransactionPatch
+          : {
+              feesStrategy,
+            };
+
+      setTransaction(bridge.updateTransaction(transaction, { ...patch, gasOptions }));
     },
-    [setTransaction, account, parentAccount, transaction],
+    [
+      setTransaction,
+      account,
+      parentAccount,
+      transaction,
+      customStrategyTransactionPatch,
+      gasOptions,
+    ],
   );
 
   const openCustomFees = useCallback(() => {
@@ -81,8 +104,13 @@ export default function EvmFeesStrategy({
       currentNavigation: ScreenName.SendSummary,
       nextNavigation: ScreenName.SendSelectDevice,
       setTransaction,
+      setCustomStrategyTransactionPatch,
     });
   }, [navigation, route.params, account.id, parentAccount, transaction, setTransaction]);
+
+  if (loading) {
+    return <InfiniteLoader size={32} />;
+  }
 
   /**
    * If no gasOptions available, this means this currency does not have a
@@ -96,7 +124,10 @@ export default function EvmFeesStrategy({
     <SelectFeesStrategy
       {...props}
       gasOptions={gasOptions}
-      customFees={customStrategy}
+      customFees={getCustomStrategyFees({
+        ...transaction,
+        ...customStrategyTransactionPatch,
+      } as Transaction)}
       onStrategySelect={onFeesSelected}
       onCustomFeesPress={openCustomFees}
       account={account}