+++ libs/ledger-live-common/src/families/cosmos/chain/chain.ts
@@ -14,6 +14,7 @@ import Stride from "./Stride";
 import Umee from "./Umee";
 import BinanceBeaconChain from "./BinanceBeaconChain";
 import Coreum from "./Coreum";
+import Injective from "./Injective";
 
 const cosmosChainParams: { [key: string]: CosmosBase } = {};
 export default function cryptoFactory(currencyId: string): CosmosBase {
@@ -65,6 +66,9 @@ export default function cryptoFactory(currencyId: string): CosmosBase {
         break;
       case "coreum":
         cosmosChainParams[currencyId] = new Coreum();
+        break;
+      case "injective":
+        cosmosChainParams[currencyId] = new Injective();
         break;
       default:
         throw new Error(`${currencyId} is not supported`);